build: optimize Makefile with -O3 -flto, debug target, strict warnings
Release: -O3 -flto -march=native + strict flags
Debug:   make debug -> -g3 -O0 -fsanitize=address,undefined,leak
Warnings: -Wshadow -Wunused -Wformat=2 -Wundef -Wstrict-prototypes
          -Wmissing-prototypes -Winit-self -Wimplicit-fallthrough=3
          -Wcast-align -Wswitch-default -Wswitch-enum
          -Wduplicated-branches -Wduplicated-cond -Wlogical-op
          -Wstack-protector -fstrict-aliasing

Fixes for strict flags:
- Add const qualifiers to er(), ft_strcmp(), string arrays
- Add static to internal functions (missing prototypes)
- Fix sign-conversion casts for ft_substr, mlx_new_image, bm_size
- Fix misleading if-indentation with braces
- Fix return-with-expression in void functions
- Fix float-equal comparisons, 10.0f -> 10.0 double promotions
- Fix shadowed 'right' parameter name
- Bugfix: parse_tx set bm_size instead of tx_size

+++ src/render/render/three_renders.c
@@ -21,7 +21,7 @@ void	render_one(void *param)
 	data->x = data->mlx->width;
 	data->y = data->mlx->height;
 	data->god = true;
-	mlx_resize_image(data->img, data->x, data->y);
+	mlx_resize_image(data->img, (uint32_t)data->x, (uint32_t)data->y);
 	new_img = render(data, 1);
 	fill_image(data, (uint32_t *)data->img->pixels, new_img);
 	mlx_image_to_window(data->mlx, data->img, 0, 0);
@@ -42,7 +42,7 @@ void	render_fast(void *param)
 	data->y = data->mlx->height;
 	data->god = true;
 	mlx_delete_image(data->mlx, data->img);
-	data->img = mlx_new_image(data->mlx, data->x, data->y);
+	data->img = mlx_new_image(data->mlx, (uint32_t)data->x, (uint32_t)data->y);
 	new_img = console_render(data);
 	fill_image(data, (uint32_t *)data->img->pixels, new_img);
 	mlx_image_to_window(data->mlx, data->img, 0, 0);
@@ -63,7 +63,7 @@ void	update_render(void *param)
 	data->x = data->mlx->width;
 	data->y = data->mlx->height;
 	data->god = true;
-	mlx_resize_image(data->img, data->x, data->y);
+	mlx_resize_image(data->img, (uint32_t)data->x, (uint32_t)data->y);
 	if (!data->img_last)
 		data->img_last = render(data, 0);
 	if (!data->img->enabled)
